refactor: change clientModule with plugin pattern for the acknowledgement
