+++ packages/common-config/plugins/acknowledgementPlugin.ts
@@ -1,4 +1,4 @@
-import ExecutionEnvironment from '@docusaurus/ExecutionEnvironment';
+import type { LoadContext, Plugin } from "@docusaurus/types";
 
 const acknowledgement = `
 Made with ♥ by
@@ -12,12 +12,15 @@ Made with ♥ by
 Designed by Strath (strath.work)
 `;
 
-if (ExecutionEnvironment.canUseDOM) {
-  window.addEventListener('DOMContentLoaded', () => {
-    const body = document.querySelector('body');
-    if (body) {
-      const comment = document.createComment(acknowledgement);
-      body.insertBefore(comment, body.firstChild);
-    }
-  });
-}
+export default function acknowledgementPlugin(context: LoadContext): Plugin<any> {
+  return {
+    name: "acknowledgement-plugin",
+    injectHtmlTags: () => {
+      return {
+        headTags: [
+          `<!--${acknowledgement}-->`,
+        ],
+      };
+    },
+  };
+}
