+++ packages/common-config/plugins/index.ts
@@ -1,0 +1,2 @@
+export { default as llmsTxtPlugin } from "./llmsTxtPlugin";
+export { default as acknowledgementPlugin } from "./acknowledgementPlugin";
